+++ CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Addded
+- Function to validate the number input. (#4)
+- Function to validate the roman numeral input. (#3)
+
 ## [1.0.0] - 2022-07-01
 ### Added
 - Function to convert from roman numeral to number. (#1)
 - Function to convert from number to roman numeral. (#2)
 
+[Unreleased]: https://github.com/rodrigoborgesdeoliveira/Romanumber/compare/main...v1.0.0
 [1.0.0]: https://github.com/rodrigoborgesdeoliveira/Romanumber/releases/tag/v1.0.0